core: pathfinding: compute geometry
