+++ api/osrd_infra/views/pathfinding.py
@@ -1,6 +1,7 @@
 import json
 from collections import defaultdict
 from typing import Mapping
+from pydantic import Json
 
 import requests
 from django.conf import settings
@@ -196,17 +197,17 @@ def compute_path(path, request_data, owner):
 
     waypoints, step_durations = parse_steps_input(request_data["steps"], infra)
     payload = request_pathfinding({"infra": infra.pk, "waypoints": waypoints})
+    path.geographic = json.dumps(payload.pop("geo_geom"))
+    path.schematic = json.dumps(payload.pop("sch_geom"))
 
     # Post treatment
     track_map = fetch_track_sections_from_payload(infra, payload)
     payload = compute_path_payload(infra, payload, step_durations, track_map)
-    geographic, schematic = get_geojson_path(payload, track_map)
+    # geographic, schematic = get_geojson_path(payload, track_map)
 
     path.owner = owner
     path.infra = infra
     path.payload = payload.dict()
-    path.geographic = geographic
-    path.schematic = schematic
     path.curves = compute_curves(payload, track_map)
     path.slopes = compute_slopes(payload, track_map)
